requirements.txt
# Core Python packages for FF6 AI Player

# Web framework
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-CORS==4.0.0
python-socketio==5.10.0

# AI and Machine Learning
openai==1.6.1
numpy==1.24.3
scipy==1.11.4

# Image processing and OCR
Pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78

# Windows-specific (for screen capture)
pywin32==306

# Networking
requests==2.31.0
eventlet==0.33.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
playwright==1.40.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
colorama==0.4.6

# Logging and monitoring
coloredlogs==15.0.1

# Data persistence
jsonpickle==3.0.2
